+++ DoctorTelegramBot/src/main/java/org/telegrambots/doctortelegrambot/repositories/DoctorRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 @Repository
 public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
     @Query(value = "SELECT d FROM Doctor d WHERE d.personalToken.chatID = ?1 ")
-    Optional<Doctor> findDoctorByChatID(int chatID);
+    Optional<Doctor> findDoctorByChatID(long chatID);
 }
