+++ DoctorTelegramBot/src/main/java/org/telegrambots/doctortelegrambot/repositories/PatientRepository.java
@@ -1,5 +1,6 @@
 package org.telegrambots.doctortelegrambot.repositories;
 
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -10,5 +11,6 @@ import java.util.Optional;
 @Repository
 public interface PatientRepository extends JpaRepository<Patient, Integer> {
     @Query(value = "SELECT p FROM Patient p WHERE p.personalToken.chatID = ?1")
-    Optional<Patient> findPatientByChatID(int chatID);
+    Optional<Patient> findPatientByChatID(long chatID);
+
 }
